Add sort selector component
